requirements.txt
Django
psycopg2
psycopg2-binary
channels
Pillow
Redis
channels-redis
djangorestframework
django-filter
django-rest-swagger
django-cors-headers
black
django-rest-auth
django-allauth
djangorestframework-serializer-extensions
